refactor: extract DXF root block committers
